requirements.txt
Pillow>=10.4.0
PyMuPDF>=1.26.0
opencv-python>=4.10.0
numpy>=1.26.4
torch>=2.4.0
transformers>=4.45.0,<5
jiwer>=3.0.4
tqdm>=4.66.5
accelerate>=0.34.2
sentencepiece>=0.2.0
